requirements.txt
dcase-util==0.2.5
numpy==1.14.5  
tensorflow-gpu==1.10.0
matplotlib==3.0.3 
scikit-learn==0.20.0 
ruamel.yaml==0.15.71  
librosa==0.6.2 
scipy==1.2.1 
ipython==6.5.0 
